requirements.txt
langchain==0.2.14
langchain-openai==0.1.21
langchain-community==0.2.12
chromadb==0.5.5
unstructured[all-docs]==0.15.7
python-dotenv==1.0.1
pillow==10.4.0
pymupdf==1.24.9
tiktoken==0.7.0